feat: [ACI-387, ACI-393] badge templates synchronization
