+++ credentials/apps/badges/distribution/credly/credly_badges/__init__.py
@@ -4,4 +4,5 @@ Init module for credly_badges.
 
 from __future__ import unicode_literals
 
+
 __version__ = "0.0.1"
