+++ credentials/apps/badges/distribution/credly/credly_badges/apps.py
@@ -1,5 +1,5 @@
 from credentials.apps.badges.apps import BadgesAppConfig
-from credentials.apps.badges.toggles import is_badges_enabled, check_badges_enabled
+from credentials.apps.badges.toggles import check_badges_enabled, is_badges_enabled
 from credentials.apps.plugins.constants import PROJECT_TYPE, PluginSettings, PluginURLs, SettingsType
