+++ credentials/apps/badges/distribution/credly/credly_badges/urls.py
@@ -2,11 +2,16 @@
 Credly Badges routing configuration.
 """
 
+from django.urls import path
+
 from credentials.apps.badges.toggles import is_badges_enabled
+
+from .rest_api import CredlyWebhook
+
 
 urlpatterns = []
 
 if is_badges_enabled():
     urlpatterns = [
-        # Define urls here
+        path('api/webhook/', CredlyWebhook.as_view(), name='credly-webhook'),
     ]
