+++ credentials/apps/badges/distribution/credly/credly_badges/utils.py
@@ -1,7 +1,7 @@
 from django.shortcuts import get_object_or_404
 
-from .rest_api import CredlyAPIClient
-from .models import CredlyOrganization, BadgeTemplate
+from .api_client import CredlyAPIClient
+from .models import BadgeTemplate, CredlyOrganization
 
 
 def sync_badge_templates_for_organization(organization_id):
@@ -13,6 +13,8 @@ def sync_badge_templates_for_organization(organization_id):
 
     Raises:
         Http404: If organization is not found.
+
+    TODO: define and delete badge templates which was deleted on credly but still exists in our database
     """
     organization = get_object_or_404(CredlyOrganization, uuid=organization_id)
 
@@ -27,3 +29,46 @@ def sync_badge_templates_for_organization(organization_id):
                 'organization': organization,
             }
         )
+
+
+def handle_badge_template_created_event(data):
+    """
+    Create a new badge template.
+    """
+    badge_template = data.get('data', {}).get('badge_template', {})
+    owner = data.get('data', {}).get('badge_template', {}).get('owner', {})
+
+    organization = get_object_or_404(CredlyOrganization, uuid=owner.get('id'))
+
+    BadgeTemplate.objects.update_or_create(
+        uuid=badge_template.get('id'),
+        defaults={
+            'name': badge_template.get('name'),
+            'organization': organization,
+        }
+    )
+
+
+def handle_badge_template_changed_event(data):
+    """
+    Change the badge template.
+    """
+    badge_template = data.get('data', {}).get('badge_template', {})
+    owner = data.get('data', {}).get('badge_template', {}).get('owner', {})
+
+    organization = get_object_or_404(CredlyOrganization, uuid=owner.get('id'))
+
+    BadgeTemplate.objects.update_or_create(
+        uuid=badge_template.get('id'),
+        defaults={
+            'name': badge_template.get('name'),
+            'organization': organization,
+        }
+    )
+
+
+def handle_badge_template_deleted_event(data):
+    """
+    Deletes the badge template by provided uuid.
+    """
+    BadgeTemplate.objects.filter(uuid=data.get('data', {}).get('badge_template', {}).get('id')).delete()
